requirements.txt
openai
python-dotenv
langchain_community
faiss-cpu
langchain_nvidia_ai_endpoints
streamlit
pypdf